Icons: use `@loadable/component` instead of `React.lazy`
This makes it work in SSR environment like Next.js for example (even though the SSR is disabled for the icon components).
