+++ src/icons/README.md
@@ -1,4 +1,4 @@
-This is a port of awesome [**System UIcons**](https://systemuicons.com/) to be used directly in React. How does it work? We take _all_ available System UIcons and automatically convert them to React components so they can be used directly.
+This is a port of awesome [**System UIcons**](https://systemuicons.com/) to be used directly in React. How does it work? We take _all_ available System UIcons, optimize them and automatically convert them from SVG to React components so they can be used directly in React.
 
 ## Installation and Usage
 
@@ -27,8 +27,6 @@ All these icons are automatically adjusting to the surrounding font size and col
 There are many icons available and it would not be a good idea to bundle them all together. For this reason we are internally using [`React.lazy`](https://reactjs.org/docs/code-splitting.html#reactlazy). Support for lazy-loading can be enabled via [`@adeira/babel-preset-adeira`](https://github.com/adeira/babel-preset-adeira) like so:
 
 ```js
-// @flow strict
-
 module.exports = {
   presets: [
     [
