+++ src/icons/generator/README.md
@@ -1,0 +1,3 @@
+```bash
+yarn monorepo-babel-node src/icons/generator/generate.js
+```
