+++ src/icons/generator/templates/reactComponent.txt
@@ -1,0 +1,8 @@
+// @flow strict
+/* eslint-disable import/newline-after-import */
+
+import React, { type Node } from 'react';
+
+export default function %%COMPONENT_NAME%%(props: {}): Node {
+  return %%JSX%%;
+};
