+++ src/icons/index.js
@@ -1,6 +1,6 @@
 // @flow
 
-import React, { Suspense, type Node } from 'react';
+import React, { type Node } from 'react';
 
 import { ComponentsMap, type IconNames } from './__generated__/__meta';
 
@@ -8,32 +8,18 @@ type Props = {
   +name: IconNames,
 };
 
-const Fallback = () => {
-  return (
-    <span
-      style={{
-        // this creates the same empty space as any loaded icon so the UI is not glitching
-        width: '1em',
-        height: '1em',
-        display: 'inline-block',
-      }}
-    />
-  );
-};
-
 export default function Icon(props: Props): Node {
   const IconComponent = ComponentsMap[props.name];
 
   return (
-    <Suspense fallback={<Fallback />}>
-      <span
-        style={{
-          verticalAlign: 'middle',
-          display: 'inline-block',
-        }}
-      >
-        <IconComponent />
-      </span>
-    </Suspense>
+    <span
+      style={{
+        // aligns icon in the middle of the text
+        verticalAlign: 'middle',
+        display: 'inline-block',
+      }}
+    >
+      <IconComponent />
+    </span>
   );
 }
